Move PalantierFormatBaseModule into its own file
